+++ aivia_demo_src.SQLDatabase/dbo/Tables/v_log_based.sql
@@ -1,8 +1,0 @@
-CREATE TABLE [dbo].[v_log_based] (
-    [IN_OR_DTTM] DATETIME2 (7)   NULL,
-    [LOG_ID]     DECIMAL (18, 4) NULL
-);
-
-
-GO
-